+++ src/components/admin/scriptures/ScriptureList.js
@@ -3,7 +3,8 @@ import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { ITEM_PER_PAGE } from '../../../utils/types';
 import Pagination from '../../shared/Pagination';
-import { getScriptures, deleteScripture } from '../../../actions/admin/admin_scriptures';
+import { getScriptures, deleteScripture } from '../../../slices/admin/adminScriptureSlice';
+import { confirmBeforeDeletion } from '../../../utils/utilityFunctions';
 
 const ScriptureList = () => {
   const dispatch = useDispatch();
@@ -48,7 +49,7 @@ const ScriptureList = () => {
   }
 
   const deleteToScripture = (id) => {
-    dispatch(deleteScripture(id));
+    if(confirmBeforeDeletion()){ dispatch(deleteScripture(id));}
   }
 
   const getAddArticleUrl = (scripture) => {
